mailing modified

+++ laravel_api/database/migrations/2024_07_15_081520_2014_10_12_100000_create_users_table.php.php
@@ -20,13 +20,18 @@ return new class extends Migration
             $table->string('email')->unique();
             $table->string('dob');
             $table->timestamp('email_verified_at')->nullable();
-            $table->integer('role_id');
+            $table->unsignedBigInteger('role_id'); // Remove ->change()
             $table->string('username')->unique();
             $table->string('password');
             $table->string('image');
+            
+            // Comment out the foreign key constraint for now
+            // $table->foreign('role_id')->references('id')->on('user_roles')->onDelete('cascade');
+            
             $table->rememberToken();
             $table->timestamps();
         });
+
     }
 
     /**
